refactor: implement class merging
