+++ lib/interfaces/graph.interface.ts
@@ -1,13 +1,13 @@
 import { IPredicate, IUpdater, IVertex, uuidType } from './params.interface';
 
-export abstract class FileGraphAbstract {
+export interface IFileGraph {
   /**
    * Creates a node in the graph.
    * @template T The date type from the node, which must match the type of the object.
    * @param {T} data - The data vertex for creating the node.
    * @returns {IVertex<T>} - created node.
    */
-  public abstract createVertex<T extends object>(data: T): Promise<IVertex<T>>;
+  createVertex<T extends object>(data: T): Promise<IVertex<T>>;
 
   /**
    * Creates multiple vertices in the graph.
@@ -17,9 +17,7 @@ export abstract class FileGraphAbstract {
    * @param {T[]} data - An array of data to be created.
    * @returns {Promise<uuidType[]>} A promise that resolves to an array of unique identifiers of the created vertices.
    */
-  public abstract createVertices<T extends object>(
-    data: T[],
-  ): Promise<IVertex<T>[]>;
+  createVertices<T extends object>(data: T[]): Promise<IVertex<T>[]>;
 
   /**
    * Updates a vertex record in the storage that matches the given condition.
@@ -32,9 +30,7 @@ export abstract class FileGraphAbstract {
    * @param {Function} updater A function that takes a vertex object {@link  IVertex<T>} and returns an updated vertex object {@link  IVertex<T>}.
    * @returns {Promise<boolean>} A promise that resolves to `true` if the record was successfully updated, and `false` if the record was not found or an error occurred.
    */
-  public abstract updateVertex<T extends object>(
-    updater: IUpdater<T>,
-  ): Promise<boolean>;
+  updateVertex<T extends object>(updater: IUpdater<T>): Promise<boolean>;
 
   /**
    * Deletes a vertex record in the storage that matches the given condition.
@@ -47,9 +43,7 @@ export abstract class FileGraphAbstract {
    * @param {Function} predicate A function that takes a vertex object {@link T & IVertex} and returns `true` if the vertex should be deleted.
    * @returns {Promise<boolean>} A promise that resolves to `true` if the record was successfully deleted, and `false` if the record was not found or an error occurred.
    */
-  public abstract deleteVertex<T extends object>(
-    predicate: IPredicate<T>,
-  ): Promise<boolean>;
+  deleteVertex<T extends object>(predicate: IPredicate<T>): Promise<boolean>;
 
   /**
    * Searches for a vertex in the storage file that matches the given predicate.
@@ -57,7 +51,7 @@ export abstract class FileGraphAbstract {
    * @param {function(T): boolean} predicate - A function that takes a vertex as input and returns `true` if the vertex matches the search criteria, otherwise `false`.
    * @returns {Promise<T | null>} A promise that resolves to the found vertex if a match is found, or `null` if no match is found.
    */
-  public abstract findOne<T extends object>(
+  findOne<T extends object>(
     predicate: IPredicate<T>,
   ): Promise<IVertex<T> | null>;
 
@@ -70,9 +64,7 @@ export abstract class FileGraphAbstract {
    * @param {function(T): boolean} predicate - A function that takes a vertex as input and returns `true` if the vertex matches the search criteria, otherwise `false`.
    * @returns {Promise<T[]>} A promise that resolves to an array of vertices that match the search criteria.
    */
-  public abstract findAll<T extends object>(
-    predicate: IPredicate<T>,
-  ): Promise<IVertex<T>[]>;
+  findAll<T extends object>(predicate: IPredicate<T>): Promise<IVertex<T>[]>;
 
   /**
    * Creates edges (links) between the specified vertices.
@@ -85,7 +77,7 @@ export abstract class FileGraphAbstract {
    *
    * @returns {Promise<boolean>} A promise that resolves to a boolean indicating whether the edge creation was successful.
    */
-  public abstract createEdge(ids: uuidType[]): Promise<boolean>;
+  createEdge(ids: uuidType[]): Promise<boolean>;
 
   /**
    * Creates an arc (edge) from one vertex to another in the graph.
@@ -99,7 +91,7 @@ export abstract class FileGraphAbstract {
    *                              or `false` if the arc already exists.
    * @throws {Error} Throws an error if the target vertex does not exist or if an error occurs during the update.
    */
-  public abstract createArc(
+  createArc(
     sourceVertexId: uuidType,
     targetVertexId: uuidType,
   ): Promise<boolean>;
@@ -116,7 +108,7 @@ export abstract class FileGraphAbstract {
    *                              or `false` if the arc does not exist.
    * @throws {Error} Throws an error if the target vertex does not exist or if an error occurs during the update.
    */
-  public abstract removeArc(
+  removeArc(
     sourceVertexId: uuidType,
     targetVertexId: uuidType,
   ): Promise<boolean>;
@@ -131,8 +123,5 @@ export abstract class FileGraphAbstract {
    * @param {uuidType} targetVertexId - The ID of the vertex to which the arc points.
    * @returns {Promise<boolean>} A promise that resolves to `true` if the arc exists, `false` otherwise.
    */
-  public abstract hasArc(
-    sourceVertexId: uuidType,
-    targetVertexId: uuidType,
-  ): Promise<boolean>;
+  hasArc(sourceVertexId: uuidType, targetVertexId: uuidType): Promise<boolean>;
 }
